Migrating FieldList to Context wrapped components
