+++ graylog2-web-interface/package.json
@@ -13,7 +13,7 @@
     "docs:build": "styleguidist build",
     "docs:server": "styleguidist server",
     "flow-check": "flow check",
-    "start": "node devServer.js",
+    "start": "FEATURES=search_3_2 node devServer.js",
     "build": "disable_plugins=true webpack --config webpack.bundled.js",
     "build:stats": "disable_plugins=true webpack --profile --json > ./webpack/stats-$(git rev-parse --abbrev-ref HEAD).json",
     "build:analyze": "yarn build:stats && webpack-bundle-analyzer ./webpack/stats-$(git rev-parse --abbrev-ref HEAD).json ./target/web/build --mode=\"static\" --report=\"./webpack/analysis-$(git rev-parse --abbrev-ref HEAD).html\"",
