+++ graylog2-web-interface/src/views/components/sidebar/fieldList/FieldList.css
@@ -16,3 +16,7 @@
     flex-grow: 1;
     margin-right: 5px;
 }
+
+:local(.filterInputContainer) input {
+    width: 100%;
+}
