+++ graylog2-web-interface/src/views/components/sidebar/index.jsx
@@ -1,4 +1,4 @@
 export { default as AddWidgetButton } from './AddWidgetButton';
-export { default as FieldList } from './FieldList';
+export { default as FieldList } from './fieldList/FieldList';
 export { default as SearchResultOverview } from './SearchResultOverview';
 export { default as SideBar } from './SideBar';
